+++ backend/app/api/v1/notes/__init__.py
@@ -1,6 +1,6 @@
 from app.api_controller import route as rt
 from app.api_controller import Access
-from app.access_managment import authorized_user, authorized_group, authorized
+from app.access_management import authorized_user, authorized_group, authorized
 from flask import request
 from . import note
